# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hooklane API",
    "version": "1.0.0",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "description": "Single-team, self-hosted webhook delivery and replay. All management routes require the administrator bearer token or a signed browser session; only event ingestion accepts the optional ingest-only bearer token. Cookie-authenticated mutations must send an Origin exactly matching the public application origin. All /api/ responses use Cache-Control: no-store. JSON input objects reject unknown fields. API operations have an eight-second server context deadline. Responses contain metadata and safe error codes, never payloads, receiver response bodies, tokens, or signing secrets. UTF-8 byte limits described in schemas are authoritative when they differ from JSON Schema character counts."
  },
  "servers": [
    {
      "url": "/",
      "description": "Same-origin self-hosted installation."
    }
  ],
...
  "security": [
    {
      "AdminBearer": []
    },
    {
      "SessionCookie": []
    }
  ],
  "paths": {
...
    "/api/v1/session": {
      "post": {
        "operationId": "createSession",
        "summary": "Sign in to the management UI",
        "tags": [
          "Session"
        ],
        "responses": {
          "200": {
            "description": "Authenticated. Sets a signed, HttpOnly, SameSite=Strict cookie valid for 12 hours.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            },
            "headers": {
              "Set-Cookie": {
                "schema": {
                  "type": "string"
                },
                "description": "hooklane_session cookie. Secure is enabled when configured for HTTPS."
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "413": {
            "$ref": "#/components/responses/Error413"
          },
          "415": {
            "$ref": "#/components/responses/Error415"
          },
          "429": {
            "$ref": "#/components/responses/Error429"
          },
          "503": {
            "$ref": "#/components/responses/Error503"
          }
        },
        "description": "Request body is limited to 4096 bytes. The per-process login limiter allows a burst of 10 attempts and replenishes one attempt every six seconds. When an Origin header is supplied, it must match the application origin. Session endpoints remain available while database initialization retries.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Login"
              }
            }
          }
        },
        "security": []
      },
      "get": {
        "operationId": "getSession",
        "summary": "Check the current management session",
        "tags": [
          "Session"
        ],
        "responses": {
          "200": {
            "description": "Authenticated with an administrator bearer token or valid session cookie.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "503": {
            "$ref": "#/components/responses/Error503"
          }
        }
      },
      "delete": {
        "operationId": "deleteSession",
        "summary": "Clear the browser session cookie",
        "tags": [
          "Session"
        ],
        "responses": {
          "204": {
            "description": "Session cookie expired in the response."
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "503": {
            "$ref": "#/components/responses/Error503"
          }
        },
        "description": "Cookie-authenticated requests require a same-origin Origin header. This clears the current browser cookie; it does not revoke previously copied signed cookies. Administrator-token rotation invalidates every session."
      }
    },
    "/api/v1/destinations": {
      "post": {
        "operationId": "createDestination",
        "summary": "Create a delivery destination",
        "tags": [
          "Destinations"
        ],
        "responses": {
          "201": {
            "description": "Destination created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Destination"
                }
              }
            },
            "headers": {
              "Location": {
                "schema": {
                  "type": "string"
                },
                "description": "API path of the created destination."
              },
              "ETag": {
                "description": "Quoted current destination revision.",
                "schema": {
                  "type": "string"
                },
                "example": "\"2\""
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "413": {
            "$ref": "#/components/responses/Error413"
          },
          "415": {
            "$ref": "#/components/responses/Error415"
          },
          "503": {
            "$ref": "#/components/responses/Error503"
          }
        },
        "description": "The request is capped at 16 KiB. Destination DNS and network policy are validated before acceptance and again on every outbound connection.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDestination"
              }
            }
          }
        }
      },
      "get": {
        "operationId": "listDestinations",
        "summary": "List destinations",
        "tags": [
          "Destinations"
        ],
        "responses": {
          "200": {
            "description": "Destinations in descending identifier order, including archived records.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DestinationPage"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "503": {
            "$ref": "#/components/responses/Error503"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/Before"
          }
        ]
      }
    },
...
    "/api/v1/events/{id}": {
      "get": {
        "operationId": "getEvent",
        "summary": "Inspect an event and its recent delivery lineage",
        "tags": [
          "Events"
        ],
        "responses": {
          "200": {
            "description": "Event metadata and up to 100 newest deliveries.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventDetail"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "503": {
            "$ref": "#/components/responses/Error503"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/ResourceID"
          }
        ]
      }
    },
    "/api/v1/events/{id}/payload": {
      "delete": {
        "operationId": "redactEventPayload",
        "summary": "Permanently redact a retained event payload",
        "tags": [
          "Events"
        ],
        "responses": {
          "204": {
            "description": "Payload removed; event metadata, digest, deliveries, and attempts preserved."
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "409": {
            "$ref": "#/components/responses/Error409"
          },
          "503": {
            "$ref": "#/components/responses/Error503"
          }
        },
        "description": "Allowed only when all deliveries for the event are terminal (succeeded, dead, or canceled). Repeated redaction succeeds. New replay requests fail once the payload has been redacted.",
        "parameters": [
          {
            "$ref": "#/components/parameters/ResourceID"
          }
        ]
      }
    },
...
    "/api/v1/stats": {
      "get": {
        "operationId": "getStats",
        "summary": "Read retained installation counts",
        "tags": [
          "Operations"
        ],
        "responses": {
          "200": {
            "description": "Destination count excludes archived destinations; event and delivery counts cover retained records only.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Stats"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "503": {
            "$ref": "#/components/responses/Error503"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "AdminBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Installation ADMIN_TOKEN. Full management and ingestion access."
      },
      "IngestBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Optional installation INGEST_TOKEN. Only POST /api/v1/events accepts this token."
      },
      "SessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "hooklane_session",
        "description": "Signed HttpOnly, SameSite=Strict cookie from POST /api/v1/session. Valid for 12 hours. Cookie mutations additionally require the matching Origin header."
      }
    },
...
    "schemas": {
...
      "Session": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "authenticated": {
            "type": "boolean",
            "const": true
          }
        },
        "required": [
          "authenticated"
        ]
      },
      "Login": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "token": {
            "type": "string",
            "writeOnly": true,
            "description": "Installation administrator token. The ingest-only token cannot create a management session."
          }
        },
        "required": [
          "token"
        ]
      },
...
    },
    "responses": {
      "Error400": {
        "description": "Invalid JSON, resource identifier, input, filter, or idempotency key.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Error401": {
        "description": "Administrator authentication is required; only ingestion also accepts the configured ingest token.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "headers": {
          "WWW-Authenticate": {
            "schema": {
              "type": "string"
            },
            "description": "Bearer realm=\"hooklane\""
          }
        }
      },
      "Error403": {
        "description": "Missing or mismatched Origin on a cookie-authenticated mutation, or mismatched Origin on login.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "Error413": {
        "description": "Payload or request body exceeds the configured endpoint limit.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Error415": {
        "description": "Content-Type must be application/json for endpoints with a JSON request body.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    }
  }
}